Add external_function and call helpers to _air_ops_ext.py
